Refactor metadata.py file location and update imports
The file metadata.py has been relocated out of the base directory and moved one level up, resulting in changes to all the import paths that referenced it. The restructuring simplifies the module's directory tree and enhances the accessibility of metadata.py.
