+++ src/mdio/schemas/metadata.py
@@ -6,7 +6,18 @@ from typing import Any
 
 from pydantic import Field
 
+from mdio.schemas.chunk_grid import RectilinearChunkGrid
+from mdio.schemas.chunk_grid import RegularChunkGrid
 from mdio.schemas.core import StrictCamelBaseModel
+
+
+class ChunkGridMetadata(StrictCamelBaseModel):
+    """Definition of chunk grid."""
+
+    chunk_grid: RegularChunkGrid | RectilinearChunkGrid | None = Field(
+        default=None,
+        description="Chunk grid specification for the array.",
+    )
 
 
 def follows_metadata_key_convention(key: str) -> bool:
