+++ src/mdio/schemas/v0/stats.py
@@ -10,8 +10,8 @@
 
 from pydantic import Field
 
-from mdio.schemas.base.metadata import VersionedMetadataConvention
 from mdio.schemas.core import StrictCamelBaseModel
+from mdio.schemas.metadata import VersionedMetadataConvention
 
 
 class Statistics(StrictCamelBaseModel):
