+++ src/mdio/schemas/v0/variable.py
@@ -7,10 +7,10 @@ from pydantic import create_model
 from mdio.schemas import NamedArray
 from mdio.schemas import ScalarType
 from mdio.schemas import StructuredType
-from mdio.schemas.base.encoding import ChunkGridMetadata
-from mdio.schemas.base.metadata import MetadataContainer
-from mdio.schemas.base.metadata import UserAttributes
 from mdio.schemas.core import model_fields
+from mdio.schemas.metadata import ChunkGridMetadata
+from mdio.schemas.metadata import MetadataContainer
+from mdio.schemas.metadata import UserAttributes
 
 
 VariableMetadata = create_model(
