+++ src/mdio/schemas/v1/dataset.py
@@ -3,9 +3,9 @@
 
 from pydantic import Field
 
-from mdio.schemas.base.metadata import UserAttributes
 from mdio.schemas.dataset import BaseDataset
 from mdio.schemas.dimension import NamedDimension
+from mdio.schemas.metadata import UserAttributes
 from mdio.schemas.v1.variable import Variable
